Separate guardian app into multiple modules
